Rewrite filesystem to fix Heroku bug
-Backend is now considerably smaller, but no official production
 functionality has been added.
