+++ .eslintrc.json
@@ -4,7 +4,7 @@
         "es6": true,
         "node": true
     },
-    "extends": ["eslint:recommended"],
+    "extends": ["eslint:recommended", "react-app"],
     "parserOptions": {
         "sourceType": "module"
     },
