--- client/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "react-ui",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "bootstrap": "^3.3.7",
-    "fetch": "^1.1.0",
-    "node-fetch": "^1.7.1",
-    "process-nextick-args": "^1.0.7",
-    "react": "^15.6.1",
-    "react-bootstrap": "^0.31.0",
-    "react-dom": "^15.6.1",
-    "react-scripts": "1.0.7",
-    "util-deprecate": "^1.0.2"
-  },
-  "devDependencies": {
-    "chai": "^4.0.2",
-    "eslint": "^4.0.0",
-    "sinon": "^2.3.5",
-    "wdio-mocha-framework": "^0.5.10",
-    "webdriverio": "^4.8.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  },
-  "proxy": "http://localhost:3001"
-}
