--- client/src/index.css
@@ -1,9 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-}
-
-thead tr td{
-  font-weight: bold;
-}
