--- client/webdriverio-test/test.js
@@ -1,14 +1,0 @@
-var webdriverio = require('webdriverio');
-var options = {
-    desiredCapabilities: {
-        browserName: 'firefox'
-    }
-};
-webdriverio
-    .remote(options)
-    .init()
-    .url('http://www.google.com')
-    .getTitle().then(function(title) {
-        console.log('Title was: ' + title);
-    })
-    .end();
