+++ react-ui/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "react-ui",
+  "version": "0.1.0",
+  "private": true,
+  "devDependencies": {
+    "chai": "^4.0.2",
+    "react-scripts": "0.9.5",
+    "wdio-mocha-framework": "^0.5.10",
+    "webdriverio": "^4.8.0"
+  },
+  "dependencies": {
+    "bootstrap": "^3.3.7",
+    "react": "^15.4.2",
+    "react-bootstrap": "^0.31.0",
+    "react-dom": "^15.4.2"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
+  },
+  "proxy": "http://localhost:5000"
+}
