+++ react-ui/src/index.css
@@ -1,0 +1,9 @@
+body {
+  margin: 0;
+  padding: 0;
+  font-family: sans-serif;
+}
+
+thead tr td {
+  font-weight: bold;
+}
